chore: Add Button component for creating PDF
The commit adds a new Button component in the `Button.jsx` file. This component is responsible for rendering a button that triggers the creation of a PDF. The Button component accepts an `onClick` prop to handle the button click event. This change is necessary to provide a user-friendly way to generate PDFs in the application.