feat: Add guest and companion columns to invitationCodes table
This commit modifies the `db.sql` file to add two new columns, `guest` and `companion`, to the `invitationCodes` table. These columns will store the names of the guest and their companion for each invitation code. This change is necessary to capture additional information about the guests and their companions in the application.